implementata interfaccia Proprietario

+++ src/main/java/it/unina/vetbook/dto/AnimaleDTO.java
@@ -1,14 +1,14 @@
 package it.unina.vetbook.dto;
 
 public class AnimaleDTO {
-    public final int codiceChip;
+    public final long codiceChip;
     public final String nome;
     public final String tipo;
     public final String razza;
     public final String colore;
     public final String dataNascita;
 
-    public AnimaleDTO(int codiceChip, String nome, String tipo, String razza, String colore, String dataNascita) {
+    public AnimaleDTO(long codiceChip, String nome, String tipo, String razza, String colore, String dataNascita) {
         this.codiceChip = codiceChip;
         this.nome = nome;
         this.tipo = tipo;
